+++ web/public/js/app.js
@@ -72,7 +72,127 @@ function displayToast (message) {
   initToastButton()
 }
 
+class Cookies {
+  constructor () {
+    this.rawCookies = document.cookie
+    this.cookies = new Map()
+
+    this.parseCookies()
+  }
+
+  getCookie (cname) {
+    return this.cookies.get(cname)
+  }
+
+  setCookie (cname, cvalue) {
+    const today = new Date()
+    const nextMonth = today.setMonth(today.getMonth() + 1)
+
+    document.cookie = `${cname}=${JSON.stringify(cvalue)};expires=${new Date(nextMonth).toUTCString()};path=/`
+
+    this.parseCookies()
+  }
+
+  parseCookies () {
+    const cookies = this.rawCookies.split('; ').filter(string => string.length)
+
+    for (let index = 0; index < cookies.length; index++) {
+      const cookieStringArr = cookies[index].split('=')
+      const cookieName = cookieStringArr[0].trim()
+      const cookieValue = cookieStringArr[1].trim()
+
+      this.cookies.set(cookieName, JSON.parse(cookieValue))
+    }
+  }
+}
+
+function setToSCookies () {
+  const login = document.getElementById('login')
+
+  if (!(login instanceof window.HTMLElement)) {
+    return
+  }
+
+  const cookies = new Cookies()
+
+  const path = window.location.pathname
+  switch (path) {
+    case '/': {
+      if (!cookies.getCookie('privacyPolicyAccepted')) {
+        login.href = '/privacy'
+      } else if (!cookies.getCookie('termsOfServiceAccepted')) {
+        login.href = '/terms'
+      }
+
+      if (!login.href.startsWith('https://id')) {
+        const el = login.lastElementChild
+        if (el) {
+          el.textContent = 'Register'
+        }
+      }
+      break
+    }
+    case '/privacy': {
+      login.firstElementChild.firstElementChild.addEventListener('click', ev => {
+        ev.preventDefault()
+
+        cookies.setCookie('privacyPolicyAccepted', true)
+
+        displayToast({ message: 'Successfully accepted privacy policy, please wait a second...' })
+
+        document.body.firstElementChild.classList.add('blur')
+
+        setTimeout(() => {
+          window.location.href = login.firstElementChild.firstElementChild.href
+        }, 2500)
+      })
+      break
+    }
+    case '/terms': {
+      if (!cookies.getCookie('privacyPolicyAccepted')) {
+        login.lastElementChild.href = window.location.origin + '/privacy'
+        login.lastElementChild.href = window.location.origin + 'Privacy policy'
+        login.lastElementChild.addEventListener('click', ev => {
+          ev.preventDefault()
+
+          displayToast({
+            error: 'Error: You have not accepted Privacy Policy yet.\n' +
+            'Please accept Privacy Policy first. Redirecting you...',
+          })
+
+          setTimeout(() => {
+            window.location.href = login.lastElementChild.href
+          }, 2500)
+        })
+        return
+      }
+
+      if (cookies.getCookie('termsOfServiceAccepted')) {
+        login.innerHTML = ''
+        login.appendChild(html.node`<p>You've accepted the Terms of Service already.</p>`)
+        return
+      }
+
+      login.lastElementChild.addEventListener('click', ev => {
+        ev.preventDefault()
+
+        cookies.setCookie('termsOfServiceAccepted', true)
+
+        displayToast({ message: 'Successfully accepted Terms of Service, please wait a second...' })
+
+        document.body.firstElementChild.classList.add('blur')
+
+        setTimeout(() => {
+          window.location.href = login.lastElementChild.href
+        }, 2500)
+      })
+      break
+    }
+  }
+}
+
 initBlurButtons()
 initToastButton()
+setToSCookies()
 
 console.log('Loaded.')